--- a/rust_con/Cargo.toml
+++ b/rust_con/Cargo.toml
@@ -14,7 +14,7 @@ serde = { version = "1.0.189", features = ["derive"] }
 serde_json = "1.0.107"
 
 [profile.release]
-lto = "fat"
+lto = "thin"
 codegen-units = 1
 panic = "abort"
 debug = false