Use least_n as an iterator

break up the single function into a main which sets up the runtime (similar to other languages which do this automatically), and allows us to only time the algorithm.

Break the types and generic re-used code into files and functions, with the intention on having this code shared between rust_con and rust solutions