@jinyus 2023-10-18
Main branch:
```
13.59 ms
179.47 ms
1.58 s
```

This PR:
```
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 12.811808ms
        Processing time (w/o IO): 12.227598ms
        Processing time (w/o IO): 12.167097ms
        Processing time (w/o IO): 12.308499ms
        Processing time (w/o IO): 12.210598ms
        Processing time (w/o IO): 12.215098ms
        Processing time (w/o IO): 12.228998ms
        Processing time (w/o IO): 12.145997ms
        Processing time (w/o IO): 22.030956ms
        Processing time (w/o IO): 12.226798ms
        Processing time (w/o IO): 12.192697ms
        Processing time (w/o IO): 12.180497ms
        Processing time (w/o IO): 12.288699ms
          Time (mean ± σ):      30.7 ms ±   3.4 ms    [User: 32.1 ms, System: 11.4 ms]
          Range (min … max):    28.9 ms …  40.5 ms    10 runs   

Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 174.790632ms
        Processing time (w/o IO): 173.682013ms
        Processing time (w/o IO): 175.778056ms
          Time (mean ± σ):     238.1 ms ±   0.7 ms    [User: 380.0 ms, System: 29.3 ms]
          Range (min … max):   237.6 ms … 238.6 ms    2 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 1.517630208s
        Processing time (w/o IO): 1.525784242s
        Processing time (w/o IO): 1.516177985s
          Time (mean ± σ):      1.738 s ±  0.013 s    [User: 3.161 s, System: 0.071 s]
          Range (min … max):    1.729 s …  1.748 s    2 runs
```